# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scrap Payments Manager API",
        "description": "API documentation for Scrap Payments Manager application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/origins": {
            "get": {
                "tags": [
                    "Origins"
                ],
                "summary": "Listar todos los lugares de procedencia",
                "operationId": "4bda38e7bab81551720103c6ab346978",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Origins"
                ],
                "summary": "Crear un nuevo lugar de procedencia",
                "operationId": "5fba01905f8e71bf062f5d59a24589eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lima"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ciudad de Lima",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recurso creado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/origins/{id}": {
            "get": {
                "tags": [
                    "Origins"
                ],
                "summary": "Obtener detalles de un lugar de procedencia específico",
                "operationId": "a8691735692b21232f7f50801c26f03e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar de procedencia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Origins"
                ],
                "summary": "Actualizar un lugar de procedencia existente",
                "operationId": "b2f220593f979cd5d4b6e7facfcee3d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar de procedencia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado exitosamente"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Origins"
                ],
                "summary": "Eliminar un lugar de procedencia",
                "operationId": "fc844f18fe17c8d032389d7572154852",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar de procedencia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso eliminado exitosamente"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}